+++ CHANGELOG.md
@@ -13,11 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Methods:
   - `isEmpty` | `isNotEmpty`
+  - `isBlank` | `isNotBlank`
 
 ### ⛏️ Work in progress
 
 - Methods:
-  - `isBlank` | `isNotBlank`
   - `isAlphaDigit` | `isNotAlphaDigit`
 
 ### 💡 Ideas for future release